convert to package :package:
